# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LeSGo API",
        "description": "Logistics & multi-service API documentation for LeSGo.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/wallets/{user_id}": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get wallet balance for a user (owner/admin only)",
                "description": "Get wallet by user ID (secured).",
                "operationId": "07d01a7b7e886c5010f910a486d803ad",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wallet"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/wallets/{user_id}/transactions": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "List wallet transactions for a user (owner/admin only)",
                "description": "List wallet transactions for a user (secured).",
                "operationId": "681f518592b42481ed7479430ba0b856",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by transaction type (credit/debit)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "credit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of wallet transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WalletTransaction"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Wallet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "balance": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "currency": {
                        "type": "string",
                        "example": "PHP"
                    }
                },
                "type": "object"
            },
            "WalletTransaction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "wallet_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "credit"
                    },
                    "source_type": {
                        "type": "string",
                        "example": "order"
                    },
                    "source_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "balance_before": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "balance_after": {
                        "type": "number",
                        "format": "float",
                        "example": 350
                    },
                    "description": {
                        "type": "string",
                        "example": "Earnings from completed trip"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}